✂️ split chart into decoupled component
